requirements.txt
numpy==2.2.5
opencv-python-headless
Pillow==11.2.1
streamlit==1.45.0
ultralytics==8.3.124
